% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisplot.R
\name{axisplot}
\alias{axisplot}
\title{Adds the y = 0 and x = 0 lines to an existing plot}
\usage{
axisplot(xy = FALSE)
}
\arguments{
\item{xy}{logical (default FALSE): plot y = x line?}
}
\description{
Adds the y = 0 and x = 0 lines to an existing plot
}
\examples{
plot(rnorm(100), rnorm(100))
axisplot()

}
\author{
N. F. Grinberg, \email{ng414@medschl.cam.ac.uk}
}

