% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_write.R
\name{blank.write}
\alias{blank.write}
\title{Write a blank line to a file; useful for creating a script log}
\usage{
blank.write(file, append = TRUE)
}
\arguments{
\item{file}{Path to a file}

\item{append}{TRUE if would like to add to file rather than overwrite it or create a new one}
}
\description{
Write a blank line to a file; useful for creating a script log
}
\examples{
# Append a blank line to log_file.txt 
blank.write(file = 'log_file.txt', append = FALSE)

}
\author{
N. F. Grinberg, \email{ng414@medschl.cam.ac.uk}
}

