% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timed_write.R
\name{timed.write}
\alias{timed.write}
\title{Write a message with a time stamp to a file; useful for creating a script log}
\usage{
timed.write(x, file, append = TRUE)
}
\arguments{
\item{x}{Anything  you would like to write to a file}

\item{file}{Path to a file where to write x}

\item{append}{TRUE if would like to add to file rather than overwrite it or create a new one}
}
\description{
Write a message with a time stamp to a file; useful for creating a script log
}
\examples{
# Write "Start: <time now>" to log_file.txt 
# (overwrite if it exists or create if it doesn't)
timed.write('Start', file = 'log_file.txt', append = FALSE)

# Append "Done: <time now>" to log_file.txt
file.create('log_file.txt')
timed.write('Done', file = 'log_file.txt')
}
\author{
N. F. Grinberg, \email{ng414@medschl.cam.ac.uk}
}
\keyword{max}

