% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyline.R
\name{xyline}
\alias{xyline}
\title{Adds the y = x plot to a figure currently in device}
\usage{
xyline()
}
\description{
Adds the y = x plot to a figure currently in device
}
\examples{
plot(rnorm(100), rnorm(100))
xyline()

}
\author{
N. F. Grinberg, \email{ng414@medschl.cam.ac.uk}
}

