% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot.R
\name{xyplot}
\alias{xyplot}
\title{Adds the y = x plot to a figure currently in device}
\usage{
xyplot()
}
\description{
Adds the y = x plot to a figure currently in device
}
\examples{
plot(rnorm(100), rnorm(100))
xyplot()

}
\author{
N. F. Grinberg, \email{ng414@medschl.cam.ac.uk}
}

