% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mse.R
\name{calc.mse}
\alias{calc.mse}
\title{Calculate mean squared error for a vector of observed and predicted values}
\usage{
calc.mse(obs, pred, rsq = FALSE)
}
\arguments{
\item{obs}{Stuff stuff}

\item{pred}{Stuff stuff stuff}

\item{rsq}{Stuff stuff stuff}
}
\value{
mse or R^2
}
\description{
Calculate mean squared error for a vector of observed and predicted values
}
\examples{

}
\author{
N. F. Grinberg, \email{ng414@medschl.cam.ac.uk}
}
\keyword{mse}

