% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxz.R
\name{maxz}
\alias{maxz}
\title{The \eqn{f(x) = max\{x, 0\}} function}
\usage{
maxz(x)
}
\arguments{
\item{x}{Any numeric object including vectors and matrices}
}
\value{
Returns argument if it's positive, or 0 otherwise
}
\description{
The \eqn{f(x) = max\{x, 0\}} function
}
\examples{
maxz(10)
maxz(-5 : 5)
maxz(matrix(-10 : 9, nrow = 4))

}
\author{
N. F. Grinberg, \email{ng414@medsch.cam.ac.uk}
}
\keyword{max}

